create helper.go and helper_test.go
